refactor: move tab components to Tab folder
