+++ src/components/EditingPanel.tsx
@@ -1,9 +1,9 @@
 import clsx from "clsx"
 import DownloadButton from "@/components/DownloadButton"
 
-import TabFilters from "@/components/TabFilters"
-import TabRatio from "@/components/TabRatio"
-import TabTemplate from "@/components/TabTemplate"
+import TabFilters from "@/components/Tab/TabFilters"
+import TabRatio from "@/components/Tab/TabRatio"
+import TabTemplate from "@/components/Tab/TabTemplate"
 
 import RatioIcon from "@/components/Icon/RatioIcon"
 import TemplateIcon from "@/components/Icon/TemplateIcon"
